Feature(columns are filtered specific to graph type)
