+++ client/components/Chart/VictoryScatterChart.js
@@ -1,8 +1,10 @@
 import React, {Component} from 'react'
-import { download } from '../utils'
+import {download} from '../../utils'
 
+import * as d3 from 'd3'
 import {
   VictoryChart,
+  VictoryScatter,
   VictoryLine,
   VictoryZoomContainer,
   VictoryBrushContainer,
@@ -14,15 +16,10 @@ import {
   VictoryVoronoiContainer
 } from 'victory'
 
-export default class VictoryLineGraph extends Component {
-  handleZoom(domain) {
-    this.setState({zoomDomain: domain})
-  }
-
+export default class VictoryScatterChart extends Component {
   render() {
-    const changeStyle = this.props.changeStyle
     let data = this.props.data
-    let keys = Object.keys(data[0])
+    const changeStyle = this.props.changeStyle
     let y = this.props.y
     let x = this.props.x
     let downloadPNG = download.bind(this)
@@ -35,7 +32,6 @@ export default class VictoryLineGraph extends Component {
             style={{parent: {maxWidth: '100%'}}}
             width={600}
             height={470}
-            // scale={{x: 'time'}}
             padding={{left: 100, right: 25, top: 35, bottom: 75}}
             containerComponent={
               <VictoryVoronoiContainer
@@ -63,10 +59,9 @@ export default class VictoryLineGraph extends Component {
               x={100}
               y={24}
             />
-            <VictoryLine
-              style={{
-                data: {stroke: this.props.color}
-              }}
+            <VictoryScatter
+              style={{data: {fill: this.props.color}}}
+              size={7}
               data={data}
               x={x}
               y={y}
@@ -103,20 +98,27 @@ export default class VictoryLineGraph extends Component {
                 }
               ]}
             />
+            <VictoryLine data={this.props.regressionLine} x={x} y={y} />
+            <VictoryLabel
+              text={this.props.title}
+              style={{
+                fontSize: 16,
+                textAnchor: 'start',
+                verticalAnchor: 'end',
+                fill: '#000000',
+                fontFamily: 'inherit',
+                fontWeight: 'bold'
+              }}
+              x={100}
+              y={24}
+            />
+
             <VictoryAxis
               label={x}
               style={{
                 axis: {stroke: '#756f6a'},
                 axisLabel: {fontSize: 12, padding: 30}
               }}
-              tickValues={data.map(datum => datum[x])}
-              tickFormat={data.map(datum => {
-                if (typeof datum[x] === 'string') {
-                  return datum[x].slice(0, 3)
-                } else {
-                  return datum[x]
-                }
-              })}
             />
             <VictoryAxis
               dependentAxis
@@ -128,18 +130,45 @@ export default class VictoryLineGraph extends Component {
             />
           </VictoryChart>
         </div>
-        <div>
+        <div id="controls">
+          {/* <p>Left Axis:</p>
+            <select
+              onChange={e => {
+                //save old y value
+
+                let newX = this.state.y
+                if (this.state.y === null) {
+                  newX = keys[1]
+                }
+
+                changeStyle(e, 'y')
+                changeStyle(newX, 'x')
+              }}
+            >
+              <option
+                key={1}
+                name="y"
+                value={keys[1]}
+                defaultValue
+              >{`${keys[1][0].toUpperCase()}${keys[1].slice(1)}`}</option>
+              <option
+                key={0}
+                name="y"
+                value={keys[0]}
+              >{`${keys[0][0].toUpperCase()}${keys[0].slice(1)}`}</option>
+            </select> */}
+
           <button onClick={() => downloadPNG(this.props.title)}>
             Download
           </button>
-          <canvas
-            id="canvas"
-            width="600"
-            height="470"
-            display="none"
-            style={{visibility: 'hidden', zIndex: -950, position: 'absolute'}}
-          />
         </div>
+        <canvas
+          id="canvas"
+          width="600"
+          height="470"
+          display="none"
+          style={{visibility: 'hidden', zIndex: -950, position: 'absolute'}}
+        />
       </div>
     )
   }
