+++ client/components/EditGraphs/CustomizeMenu.js
@@ -1,0 +1,118 @@
+import React from 'react'
+import {buildRegressionModel} from '../../utils'
+
+export const CustomizeMenu = function(props) {
+  const changeStyle = props.changeStyle
+  const graphData = props.graphData
+  const graphSelected = props.graphSelected
+
+  const filterColumn = function(data, dataType, key, i) {
+    if (typeof data[0][key] === dataType) {
+      return (
+        <option key={i} value={key}>
+          {key}
+        </option>
+      )
+    } else {
+      return ''
+    }
+  }
+  return (
+    <div id="controls">
+      <p>Choose a Dataset:</p>
+      <select name="dataId" onChange={e => changeStyle(e, 'dataId')}>
+        <option />
+        {props.data.map((elem, i) => (
+          <option key={i} value={elem.id}>
+            {elem.id}
+          </option>
+        ))}
+      </select>
+      <p>Left Axis:</p>
+      <select name="y" onChange={e => changeStyle(e, 'y')}>
+        <option />
+        {/* {Only map columns with numeric values} */}
+        {Object.keys(graphData[0]).map((key, i) => {
+          return filterColumn(graphData, 'number', key, i)
+        })}
+      </select>
+
+      <p>Bottom Axis:</p>
+      <select name="x" onChange={e => changeStyle(e, 'x')}>
+        <option />
+        {Object.keys(graphData[0]).map((key, i) => {
+          if (graphSelected === 'bar') {
+            return filterColumn(graphData, 'string', key, i)
+          } else if (graphSelected === 'scatter') {
+            return filterColumn(graphData, 'number', key, i)
+          } else {
+            return (
+              <option key={i} value={key}>
+                {key}
+              </option>
+            )
+          }
+        })}
+      </select>
+
+      <p>Color:</p>
+      <select name="color" onChange={e => changeStyle(e, 'color')}>
+        <option value="tomato">Tomato</option>
+        <option value="gold">Gold</option>
+        <option value="orange">Orange</option>
+        <option value="#f77">Salmon</option>
+        <option value="#55e">Purple</option>
+        <option value="#8af">Periwinkle</option>
+      </select>
+      {graphSelected === 'line' ? (
+        ''
+      ) : (
+        <div>
+          <p>Highlight:</p>
+          <select name="highlight" onChange={e => changeStyle(e, 'highlight')}>
+            <option value="orange">Orange</option>
+            <option value="tomato">Tomato</option>
+            <option value="gold">Gold</option>
+            <option value="#f77">Salmon</option>
+            <option value="#55e">Purple</option>
+            <option value="#8af">Periwinkle</option>
+          </select>
+        </div>
+      )}
+
+      <p>Pointer:</p>
+      <select name="tooltip" onChange={e => changeStyle(e, 'tooltip')}>
+        <option value={5}>Round edge</option>
+        <option value={0}>Square</option>
+        <option value={25}>Circle</option>
+      </select>
+
+      <p>Graph Title:</p>
+      <input value={props.title} onChange={e => changeStyle(e, 'title')} />
+      {graphSelected === 'scatter' ? (
+        <p>
+          Regression Line:{' '}
+          <input
+            type="checkbox"
+            onChange={e => {
+              changeStyle(!props.regression, 'regression')
+              console.log('x and y on state', props.x, props.y)
+              if (!props.regression) {
+                buildRegressionModel(
+                  graphData,
+                  props.x,
+                  props.y,
+                  props.changeStyle
+                )
+              } else {
+                changeStyle([], 'regressionLine')
+              }
+            }}
+          />
+        </p>
+      ) : (
+        ''
+      )}
+    </div>
+  )
+}
