+++ client/components/TestSave.js
@@ -1,17 +1,17 @@
 //Component just for testing Proof of Concept for saving interactive graphs
 import React from 'react'
-import VictoryBarChart from './VictoryBarChart'
-import { download, addComma } from '../utils'
+import VictoryBarChart from './Chart/VictoryBarChart'
+import {download, addComma} from '../utils'
 
 let savedProps = {
-    graphSelected: 'bar',
-    color: 'tomato',
-    title: 'Test Graph',
-    highlight: 'orange',
-    tooltip: '5',
-    x: 'quarter',
-    y: 'earnings',
-    dataId: 1
+  graphSelected: 'bar',
+  color: 'tomato',
+  title: 'Test Graph',
+  highlight: 'orange',
+  tooltip: '5',
+  x: 'quarter',
+  y: 'earnings',
+  dataId: 1
 }
 
 let data = [
@@ -21,10 +21,20 @@ let data = [
   {quarter: '4', earnings: 18000, items: 70000, state: 'NY'}
 ]
 
-export default function TestSave () {
+export default function TestSave() {
   return (
     <div>
-      <VictoryBarChart color={savedProps.color} title={savedProps.title} highlight={savedProps.highlight} tooltip={savedProps.tooltip} x={savedProps.x} y={savedProps.y} data={data} addComma={addComma} download={download} />
+      <VictoryBarChart
+        color={savedProps.color}
+        title={savedProps.title}
+        highlight={savedProps.highlight}
+        tooltip={savedProps.tooltip}
+        x={savedProps.x}
+        y={savedProps.y}
+        data={data}
+        addComma={addComma}
+        download={download}
+      />
     </div>
   )
 }
