+++ client/routes.js
@@ -1,13 +1,13 @@
-import React, { Component } from 'react'
-import { connect } from 'react-redux'
-import { withRouter, Route, Switch } from 'react-router-dom'
+import React, {Component} from 'react'
+import {connect} from 'react-redux'
+import {withRouter, Route, Switch} from 'react-router-dom'
 import PropTypes from 'prop-types'
-import { Login, Signup, UserHome, FileDrop } from './components'
-import { me } from './store'
+import {Login, Signup, UserHome, FileDrop} from './components'
+import {me} from './store'
 import Dashboard from './components/Dashboard'
-import VictoryBarChart from './components/VictoryBarChart'
-import VictoryLineGraph from './components/VictoryLineGraph'
-import VictoryScatterChart from './components/VictoryScatterChart'
+import VictoryBarChart from './components/Chart/VictoryBarChart'
+import VictoryLineGraph from './components/Chart/VictoryLineGraph'
+import VictoryScatterChart from './components/Chart/VictoryScatterChart'
 import HomeView from './components/Home/HomeView'
 
 import TempDashboard from './components/TempDashboard'
@@ -28,7 +28,7 @@ class Routes extends Component {
   }
 
   render() {
-    const { isLoggedIn } = this.props
+    const {isLoggedIn} = this.props
 
     return (
       <Switch>
